Fixed unit-tests and galaxy ingest cache tempdir removal check

+++ janis_core/ingestion/galaxy/containers/cache.py
@@ -18,7 +18,8 @@ def init_cache() -> ContainerCache:
     if settings.ingest.galaxy.DISABLE_CONTAINER_CACHE:
         temp = tempfile.TemporaryFile()
         cache_path = os.path.join(tempfile.gettempdir(), str(temp.name))
-        os.remove(cache_path)
+        if os.path.exists(cache_path):
+            os.remove(cache_path)
         with open(cache_path, 'w') as fp:
             fp.write('{}')
     else:
